+++ src/main/java/roomescape/service/dto/response/reservation/UserReservationResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.response;
+package roomescape.service.dto.response.reservation;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDate;
