+++ src/main/resources/static/js/reservation-mine.js
@@ -12,7 +12,7 @@ function render(data) {
   const tableBody = document.getElementById('table-body');
   tableBody.innerHTML = '';
 
-  data.forEach(item => {
+  data.reservations.forEach(item => {
     const row = tableBody.insertRow();
 
     const theme = item.theme;
